ENH: Add function create_window_from_widget in widgets.window module.
   This function encapsulates any widget inside a SiriusDialog window.
